refactor(api): move admin-member route
to src/team
