+++ api/src/team/application/admin-member/admin-member.route.js
@@ -1,6 +1,7 @@
 import Joi from 'joi';
 
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { adminMemberController } from './admin-member.controller.js';
 
 export const adminMemberRoutes = [
@@ -42,6 +43,39 @@ export const adminMemberRoutes = [
       notes: [
         "- **Cette route est restreinte aux utilisateurs ayant le droit d'accès SUPER_ADMIN**\n" +
           '- Lister les utilisateurs ayant accès à Pix Admin \n',
+      ],
+      tags: ['api', 'admin-members'],
+    },
+  },
+  {
+    method: 'PATCH',
+    path: '/api/admin/admin-members/{id}',
+    config: {
+      pre: [
+        {
+          method: (request, h) => securityPreHandlers.checkAdminMemberHasRoleSuperAdmin(request, h),
+          assign: 'hasAuthorizationToAccessAdminScope',
+        },
+      ],
+      validate: {
+        params: Joi.object({
+          id: identifiersType.adminMemberId,
+        }),
+        payload: Joi.object({
+          data: Joi.object({
+            attributes: Joi.object({
+              role: Joi.string().valid('SUPER_ADMIN', 'SUPPORT', 'METIER', 'CERTIF'),
+            }),
+          }),
+        }),
+        options: {
+          allowUnknown: true,
+        },
+      },
+      handler: (request, h) => adminMemberController.updateAdminMember(request, h),
+      notes: [
+        "- Cette route est restreinte aux utilisateurs ayant le droit d'accès SUPER_ADMIN\n" +
+          "- Elle permet de changer le rôle d'un membre Pix Admin",
       ],
       tags: ['api', 'admin-members'],
     },
